Confictok megszüntése
